decompose stdlib into modules; fix system libdevice lookup
