+++ python/triton/language/__init__.py
@@ -65,94 +65,111 @@ from ..impl.base import (
     _where,
     _xor_,
 )
-from ..impl.core import minimum, where
-
-from .stdlib import (
-    abs,
-    annotations,
+from ..impl.core import (
+    minimum,
+    where,
+)
+from .meta import (
+    _globaltimer,
+    globaltimer,
+    _clock,
+    clock,
+    debug_barrier,
+    _program_id,
+    program_id,
+    _num_programs,
+    num_programs,
+    _multiple_of,
+    multiple_of,
+    _max_contiguous,
+    max_contiguous,
+)
+from .constructors import (
     _arange,
     arange,
-    _argmax,
-    argmax,
-    _argmin,
-    argmin,
-    _atomic_add,
-    atomic_add,
-    atomic_and,
+    _zeros,
+    zeros,
+    zeros_like,
+    _cat,
+    cat,
+)
+from .quantization import (
+    _dequantize,
+    dequantize,
+)
+from .broadcasting import (
+    broadcast,
+    broadcast_to,
+)
+from .transfer import (
+    _load,
+    load,
+    _store,
+    store,
+)
+from .structure import (
+    reshape,
+    ravel,
+    swizzle2d,
+)
+from .matrix import (
+    _dot,
+    dot,
+)
+from .atomic import (
     _atomic_cas,
     atomic_cas,
+    _atomic_xchg,
+    atomic_xchg,
+    _atomic_add,
+    atomic_add,
     _atomic_max,
     atomic_max,
     _atomic_min,
     atomic_min,
+    atomic_and,
     _atomic_or,
     atomic_or,
-    _atomic_xchg,
-    atomic_xchg,
     _atomic_xor,
     atomic_xor,
-    broadcast,
-    broadcast_to,
-    _cat,
-    cat,
-    cdiv,
-    _clock,
-    clock,
-    _cos,
-    cos,
-    debug_barrier,
-    _dequantize,
-    dequantize,
-    division,
-    _dot,
-    dot,
+)
+from .math import (
+    _umulhi,
+    umulhi,
+    fdiv,
     _exp,
     exp,
-    fdiv,
-    _globaltimer,
-    globaltimer,
-    _load,
-    load,
     _log,
     log,
+    _cos,
+    cos,
+    _sin,
+    sin,
+    _sqrt,
+    sqrt,
+    abs,
+    cdiv,
+)
+from .reductions import (
+    _argmax,
+    argmax,
+    _argmin,
+    argmin,
     _max,
     max,
-    _max_contiguous,
-    max_contiguous,
     maximum,
     _min,
     min,
-    _multiple_of,
-    multiple_of,
-    _num_programs,
-    num_programs,
-    _program_id,
-    program_id,
-    ravel,
     _reduce_impl,
     _reduce_impl,
-    reshape,
     sigmoid,
-    _sin,
-    sin,
     softmax,
-    _sqrt,
-    sqrt,
-    _store,
-    store,
     _sum,
     sum,
-    swizzle2d,
-    _umulhi,
-    umulhi,
     _xor_sum,
     _xor_sum,
     xor_sum,
-    _zeros,
-    zeros,
-    zeros_like,
 )
-
 from . import random
 from .random import (
     pair_uniform_to_normal,
@@ -167,77 +184,41 @@ from .random import (
     triton,
     uint32_to_uniform_float,
 )
-
 from . import libdevice
 
 __all__ = [
-    "abs",
     "_add",
     "_and_",
-    "annotations",
     "_arange",
     "arange",
     "_argmax",
     "argmax",
     "_argmin",
     "argmin",
-    "_atomic_add",
-    "atomic_add",
-    "atomic_and",
-    "_atomic_cas",
-    "atomic_cas",
-    "_atomic_max",
-    "atomic_max",
-    "_atomic_min",
-    "atomic_min",
-    "_atomic_or",
-    "atomic_or",
-    "_atomic_xchg",
-    "atomic_xchg",
-    "_atomic_xor",
-    "atomic_xor",
     "bfloat16",
     "_binary_op_type_checking_impl",
     "_bitcast",
     "_bitwise_op_type_checking_impl",
     "block_type",
     "_bool_like",
-    "broadcast",
     "_broadcast_impl_shape",
     "_broadcast_impl_value",
-    "broadcast_to",
     "builtin",
     "_cast",
     "_cat",
     "cat",
-    "cdiv",
     "_check_ptr_type_impl",
-    "_clock",
-    "clock",
     "_computation_type_impl",
     "constexpr",
     "_constexpr_to_value",
-    "_cos",
-    "cos",
-    "debug_barrier",
-    "_dequantize",
-    "dequantize",
-    "division",
-    "_dot",
-    "dot",
     "dtype",
     "_equal",
-    "_exp",
-    "exp",
     "_fdiv",
-    "fdiv",
     "float16",
     "float32",
     "float64",
     "float8",
     "_floordiv",
-    "_globaltimer",
-    "globaltimer",
     "_greater_equal",
     "_greater_than",
     "int1",
@@ -252,34 +233,22 @@ __all__ = [
     "is_triton_tensor",
     "_less_equal",
     "_less_than",
-    "_load",
-    "load",
-    "_log",
-    "log",
     "_lshr",
     "_max",
     "max",
-    "_max_contiguous",
-    "max_contiguous",
     "maximum",
     "_min",
     "min",
     "minimum",
     "_mod",
     "_mul",
-    "_multiple_of",
-    "multiple_of",
     "_not_equal",
-    "_num_programs",
-    "num_programs",
     "_or_",
     "pair_uniform_to_normal",
     "philox",
     "philox_impl",
     "pi32_t",
     "pointer_type",
-    "_program_id",
-    "program_id",
     "rand",
     "rand4x",
     "randint",
@@ -287,42 +256,29 @@ __all__ = [
     "randn",
     "randn4x",
     "random",
-    "ravel",
-    "_reduce_impl",
     "_reduce_impl",
     "reinterpret",
     "_reshape",
-    "reshape",
     "_shl",
     "sigmoid",
-    "_sin",
-    "sin",
     "softmax",
-    "_sqrt",
-    "sqrt",
-    "_store",
-    "store",
     "_sub",
     "_sum",
     "sum",
-    "swizzle2d",
     "tensor",
     "TensorWrapper",
-    "_truediv",
     "_to_tensor",
     "triton",
+    "_truediv",
     "uint16",
     "uint32",
     "uint32_to_uniform_float",
     "uint64",
     "uint8",
-    "_umulhi",
-    "umulhi",
     "void",
     "_where",
     "where",
     "_xor_",
-    "_xor_sum",
     "_xor_sum",
     "xor_sum",
     "_zeros",
