+++ python/triton/language/random.py
@@ -1,5 +1,6 @@
 import triton
 import triton.language as tl
+import triton.language.math
 
 PHILOX_KEY_A: tl.constexpr = tl.constexpr(-1640531527)  # 0x9E3779B9
 PHILOX_KEY_B: tl.constexpr = tl.constexpr(-1150833019)  # 0xBB67AE85
@@ -18,8 +19,8 @@ def philox_impl(c0, c1, c2, c3, k0, k1, n_rounds: tl.constexpr = N_ROUNDS_DEFAUL
         A = PHILOX_ROUND_A
         B = PHILOX_ROUND_B
         _c0, _c2 = c0, c2
-        c0 = tl.umulhi(B, _c2) ^ c1 ^ k0
-        c2 = tl.umulhi(A, _c0) ^ c3 ^ k1
+        c0 = triton.language.math.umulhi(B, _c2) ^ c1 ^ k0
+        c2 = triton.language.math.umulhi(A, _c0) ^ c3 ^ k1
         c1 = B * _c2
         c3 = A * _c0
         # raise key
@@ -118,8 +119,8 @@ def pair_uniform_to_normal(u1, u2):
     """Box-Muller transform"""
     u1 = tl.maximum(1.0e-7, u1)
     th = 6.283185307179586 * u2
-    r = tl.sqrt(-2.0 * tl.log(u1))
-    return r * tl.cos(th), r * tl.sin(th)
+    r = triton.language.math.sqrt(-2.0 * triton.language.math.log(u1))
+    return r * triton.language.math.cos(th), r * triton.language.math.sin(th)
 
 
 @triton.jit
